00南哥教程目录/MCPTest-main/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
mcp==1.1.0
uvicorn==0.32.1
asyncio==3.4.3
duckduckgo-search==6.4.2
mcp-server-fetch==0.6.2
